@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "8.0.3",
+  "version": "8.0.4",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -28,8 +28,8 @@
     "sift": "16.0.1"
   },
   "devDependencies": {
-    "@babel/core": "7.23.5",
-    "@babel/preset-env": "7.23.5",
+    "@babel/core": "7.23.7",
+    "@babel/preset-env": "7.23.7",
     "@typescript-eslint/eslint-plugin": "^6.2.1",
     "@typescript-eslint/parser": "^6.2.1",
     "acquit": "1.3.0",
@@ -45,7 +45,7 @@
     "crypto-browserify": "3.12.0",
     "dotenv": "16.3.1",
     "dox": "1.0.0",
-    "eslint": "8.55.0",
+    "eslint": "8.56.0",
     "eslint-plugin-markdown": "^3.0.1",
     "eslint-plugin-mocha-no-only": "1.1.1",
     "express": "^4.18.1",
@@ -65,8 +65,8 @@
     "q": "1.5.1",
     "sinon": "17.0.1",
     "stream-browserify": "3.0.0",
-    "tsd": "0.29.0",
-    "typescript": "5.3.2",
+    "tsd": "0.30.3",
+    "typescript": "5.3.3",
     "uuid": "9.0.1",
     "webpack": "5.89.0"
   },
