@@ -1,38 +1,37 @@
-import { NextFunction,Request, Response, } from "express";
+import { NextFunction, Request, Response } from "express";
 import ErrorHandler from "../utils/ErrorHandling";
-import { request } from "http";
 
+export const ErrorMiddleware = (err: any, req: Request, res: Response, next: NextFunction) => {
+   err.statusCode = err.statusCode || 500;
+   err.message = err.message || 'Internal server error'; 
 
-export const ErrorMiddleware = (err:any,req:Request,res:Response,next:NextFunction)=>{
-   err.statusCode= err.statusCode ||  500;
-   err.message =err.message || 'Internal server error'; 
+   // Wrong MongoDB ID error
+   if (err.name === 'CastError') {
+      const message = `Resource not found. Invalid: ${err.path}`;
+      err = new ErrorHandler(message, 400);
+   }
 
-   //wrong mongo db id error
-   if(err.name === 'CastError '){
-    const message='Resource not found. Invalid: ${err.path} ';
-    err = new ErrorHandler(message, 400);
+   // Duplicate key error
+   if (err.code === 11000) {
+      const message = `Duplicate ${Object.keys(err.keyValue)} entered`;
+      err = new ErrorHandler(message, 400); 
+   }
 
-      }
-       //duplicate key error
-       if(err.code=== 11000){
-        const message ='Duplicate ${Object.keys(err.keyValue)} entered';
-    err= new ErrorHandler(message, 400); 
-    }
-    //jwt expired error
-    if(err.name==='TokenExpiredError'){
-        const message = 'Json web token is expired, again'
-        err= new ErrorHandler(message, 400);
-    }
+   // JWT expired error
+   if (err.name === 'TokenExpiredError') {
+      const message = 'JSON Web Token is expired. Please log in again.';
+      err = new ErrorHandler(message, 400);
+   }
 
-    //jwt expired error
-    if(err.name==='TokenExpiredError'){
-        const message = 'Json web token is expired, again'
-        err = new ErrorHandler(message, 400);
-    }
+   // Additional error handling can be added here
 
-res.status(err.statusCode).json({
-    success:false,
-    message:err.message,
-});
+   // Development error logging
+   if (process.env.NODE_ENV === 'development') {
+      console.error(err);
+   }
 
+   res.status(err.statusCode).json({
+      success: false,
+      message: err.message,
+   });
 };
