@@ -5,6 +5,7 @@ import { timeStamp } from "console";
 import { Mode, promises } from "fs";
 import jwt from 'jsonwebtoken';
 
+
 const emailRegexPattern: RegExp = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
 
 export interface IUser extends Document {
@@ -16,7 +17,8 @@ export interface IUser extends Document {
         url: string;
     };
     role: string;
-    isVerified: boolean;
+    isVerified: boolean; // Tracks email verification status
+    isActivated: boolean; // Tracks whether the user's account is activated
     courses: mongoose.Types.ObjectId[]; // Array of ObjectIds
     comparePassword: (password: string) => Promise<boolean>;
 }
@@ -55,6 +57,10 @@ const userSchema: Schema<IUser> = new mongoose.Schema({
         type: Boolean,
         default: false,
     },
+    isActivated: { // New field added
+        type: Boolean,
+        default: false,
+    },
     courses: [{
         type: mongoose.Schema.Types.ObjectId,
         ref: 'Course',
@@ -70,7 +76,7 @@ userSchema.pre<IUser>('save', async function(next) {
 
 // Compare password
 userSchema.methods.comparePassword = async function(enteredPassword: string): Promise<boolean> {
-    return await bcrypt.compare(enteredPassword, this.password);
+return await bcrypt.compare(enteredPassword, this.password);
 };
 
 const UserModel: Model<IUser> = mongoose.model('User', userSchema);
